requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=23.0.0
mypy>=1.5.0